Module iota_rust_sdk::types::crypto

Structs

Bls12381PrivateKey

Bls12381PublicKey

Bls12381Signature

Bn254FieldElement

CircomG1

A G1 point in BN254 serialized as a vector of three strings which is the canonical decimal representation of the projective coordinates in Fq.

CircomG2

A G2 point in BN254 serialized as a vector of three vectors each being a vector of two strings which are the canonical decimal representation of the coefficients of the projective coordinates in Fq2.

Claim

A claim consists of value and index_mod_4.

Ed25519PrivateKey

Ed25519PublicKey

Ed25519Signature

Jwk

Struct that contains info for a JWK. A list of them for different kids can be retrieved from the JWK endpoint (e.g. https://www.googleapis.com/oauth2/v3/certs). The JWK is used to verify the JWT token.

JwkId

Key to identify a JWK, consists of iss and kid.

JwtDetails

A structured of parsed JWT details, consists of kid, header, iss.

MultisigAggregatedSignature

The struct that contains signatures and public keys necessary for authenticating a Multisig.

MultisigCommittee

MultisigMember

PasskeyAuthenticator

An passkey authenticator with parsed fields. See field definition below. Can be initialized from [struct RawPasskeyAuthenticator].

PasskeyPublicKey

Secp256k1PrivateKey

Secp256k1PublicKey

Secp256k1Signature

Secp256r1PrivateKey

Secp256r1PublicKey

Secp256r1Signature

SignatureFromBytesError

ValidatorAggregatedSignature

ValidatorCommittee

ValidatorCommitteeMember

ValidatorSignature

ZkLoginAuthenticator

An zk login authenticator with all the necessary fields.

ZkLoginInputs

All inputs required for the zk login proof verification and other public inputs.

ZkLoginProof

The struct for zk login proof.

ZkLoginPublicIdentifier

A wrapper struct to retrofit in [enum PublicKey] for zkLogin. Useful to construct [struct MultiSigPublicKey].

Enums

MultisigMemberPublicKey

MultisigMemberSignature

SignatureScheme

SimpleSignature

UserSignature